serial: pl011: Use threaded interrupt for TX
We've just added threaded interrupt handling for RX, now do the same for
TX.  The RX and TX code paths are disjunct, so convert them in separate
commits for clarity.  As before, introduce a pio_tx kthread which is
woken from the interrupt handler to send data asynchronously and which
is spawned on ->startup and killed on ->shutdown.

The serial core can start and stop TX at will and that state is tracked
by the pio_tx kthread in a bool "stopped".  The thread contains a simple
state machine that can carry out certain steps when started and stopped.
This is needed when adding rs485 support so that RTS is asserted when
started and deasserted when stopped.

Obviously, these steps are needed not only in PIO mode but also in DMA
mode.  So when the serial core invokes ->start_tx, the pio_tx thread
is woken, will carry out the steps common to DMA and PIO code, then try
switching to DMA mode by calling pl011_dma_tx_start().  If that fails or
if the PL011 is not DMA-capable, the kthread will fall back to PIO
transmission.

Conversely, upon invocation of ->stop_tx by the serial core, the pio_tx
kthread is likewise woken and will take care of disabling DMA and, in an
upcoming commit, deassert RTS in rs485 mode.  In essence the pio_tx
kthread manages starting and stopping but delegates the actual
transmission to DMA if possible.

Previously there were two possibilities to switch from PIO to DMA mode:
Via pl011_dma_tx_irq() upon reception of a TX interrupt and via
pl011_dma_tx_start() upon invocation of ->start_tx by the serial core.
With threaded interrupt handling this distinction is now gone as in both
cases all that happens is a wake-up of the pio_tx kthread which then
independently does the right thing.  Thus one of the two entry points,
pl011_dma_tx_irq(), becomes obsolete and is removed.

pl011_dma_tx_start() may now be invoked with the TX interrupt enabled or
disabled and is therefore amended to disable the interrupt upon
successfully switching to DMA mode or enabling it on failure.

When transmitting in PIO mode, the FIFO fill level is tracked as with
pio_rx to avoid a register read for checking FIFO fullness.
On ->stop_tx, the TX interrupt is now deliberately left on so that the
fill level is updated.  As an additional optimization, on ->start_tx the
"Transmit FIFO empty" bit in the Flag Register is checked, in which case
the full FIFO length is available for "blind" writing.

Signed-off-by: Lukas Wunner <lukas@wunner.de>
(cherry picked from commit 6874d547bbe1600bb1ad11a3381ae80753e9f953. This
patch is required to drastically degree the rate for parity and frame
errors when the driver is used with the piControl kernel module).
Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>