tpm: do not invalidate chip ops pointer
At chip deregistration the ops pointer is set to NULL to indicate that the
chip is not functional any more. However the chip registration can happen
while the tpm and/or tpmrm file are still opened. In case of closing the
tpmrm device file the chips ops pointer might still be accessed to cleanup
all spaces (see tpm2_del_space()) which then results in a NULL pointer
dereference.
Fix this by not setting ops to NULL to indicate an unregistered chip and
thus keeping the operations pointer functional even after deregistration.
Instead use a dedicated variable "is_alive" which is set to true at chip
registration and to false as soon as it is unregistered to mark the chip
as not functional any more.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>