af_packet: Introduce egress hook
Add egress hook for AF_PACKET sockets that have the PACKET_QDISC_BYPASS
socket option set to on, which allows packets to escape without being
filtered in the egress path.

This patch only updates the AF_PACKET path, it does not update
dev_direct_xmit() so the XDP infrastructure has a chance to bypass
Netfilter.

Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>
[lukas: acquire rcu_read_lock, fix typos, rebase]
Signed-off-by: Lukas Wunner <lukas@wunner.de>
https://patchwork.kernel.org/project/netdevbpf/patch/012e6863d0103d8dda1932d56427d1b5ba2b9619.1611304190.git.lukas@wunner.de/
Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>