serial: amba-pl011: dont calculate tx drain interval based on jiffies
Currently the tx drain interval is calculated by means of the port timeout
which is provided by the serial core. However this timeout is based on
jiffies. The jiffie resolution may not be sufficient to provide values that
result in very small drain intervals, e.g. if the baud rate is 115200
bit/s or higher.
For this reason do an own calculation of the max period to drain the tx
queue which provides usec resolution. This leads to a reduction of the
calculated tx queue drain interval from 625 to 86 usec if a baud rate of
115200 bit/s is used.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>