serial: amba-pl011: use smaller period for tx queue draining
According to Lukas Wunner it may only take 19 usecs until the response of
the DIOs is received after data is requested (see the commit message of the
patch "serial: pl011: Shorten delay to deassert rs485 Transmit Enable").

The calculated max delay used for the tx queue draining is 86 usecs with a
baud rate of 115200 bit/s. To provide a better responsiveness in the
average case divide this value by 5. For the best case this reduces the
delay to 17 usecs. For the worst case however the delay has to be repeated
multiple times up to a total of 86 usecs.

So this is only meant as a workaround to reduce the probability of problems
like data corruption in case of a communication with DIOs. The real fix
should be a in the DIO firmware in the form of a minimal delay of 86 usecs
or more before a response is sent.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>