net: dsa: tag_ksz: linearize SKB before adding DSA tag
In ksz9477_xmit() skb_put() is used to add the DSA tag to the passed SKB.
However skb_put() must only be called for linear SKBs which may not be the
case if the DSA slave device inherited NETIF_F_SG from the master device.
So make sure the SKB is always linearized.

TODO: as soon as mainline patch is available replace this fix with the
mainline version

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>