net: dsa: tag_ksz: dont let the hardware process the layer 4 checksum
If the checksum calculation is offloaded to the network device (e.g due to
NETIF_F_HW_CSUM inherited from the DSA master device), the calculated
layer 4 checksum is incorrect. This is since the DSA tag which is placed
after the layer 4 data is considered as being part of the data and thus
errorneously included into the checksum calculation.
To avoid this, always calculate the layer 4 checksum in software.

TODO: check if upstream solution is available and use that instead of this
patch

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>