arm(64)/dts/connect4: Reduce SPI frequency of KSZ9897
It turned out that if the SPI frequency for accessing the KSZ9897 is too
high (more than 33MHz) the data read from the PHY registers may be
corrupted if at the same time the USB is accessed. This may be caused by
interferences between the SPI and USB signals.
However reducing the frequency to a value below or equal to 33MHz for the
SPI communication with the switch eliminates this effect.
To be on the save side choose a value of 25MHz.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>