serial: 8250: Fix define for conditional compilation
The define CONFIG_PREEMPT_RT_FULL is used to switch from non-sleepable to
sleepable gpio access in realtime kernels. However this macro does not
exist any more so the switch has not effect and also for realtime kernels
the non-sleepable gpio access is used. Fix this by using the correct define
for newer kernels.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>