dts: revpi-core: pull down sniff pin
There is no pull-down resistor for the PB_SNIFF 2a (gpio28). So the pin is
floating if no device is connected. During device detection piControl
outputs a HIGH signal for 9ms and then configures the gpio as input again.
The signal will then read as HIGH for some time after this as there
is no pull-down attached. So when piControl actually reads the pin to
determine if any module is connected to the left side it will falsely
detect a device and then start the device configuration procedure. This
results in an error due to the missing response of the falsely detected
device.

Fix this by configuring the internal (weak) pull down of the pin controller
in the dt overlay.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>