net:enc28j60: Move xmit to seperate thread
For the package transmit, after the skb was put to the private struct (
without queue), the package would be actually deliver to hardware in workqueue.
As there are mass works in the workqueue, some latency would be resulted in.

To reduce the latency, the xmit will be put in a seperate thread, which focus only
on the delivering of the package to hardware.
And the prio of the thread can be improved to further decrease the latency.

Signed-off-by: Zhi Han <z.han@kunbus.com>