pibridge-serdev: do not clear receive fifo in pibridge_send()
In pibridge_send() the receive fifo is cleared to make sure that stale data
is not intermixed with newly requested data. However Clearing the fifo
after the data request is sent to the hardware modules is unsafe: In case
that the data arrives too fast, also the new data may be erased.

Furthermore as a low level function pibridge_send() should not clear() any
data at all but only do a single task, namely send data.

For these reasons remove the concerning code from the function and instead
clear the data in the higher-level functions pibridge_req_gate_tmt() and
pibridge_req_io() which implement complete transactions.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>