pibridge-serdev: improve error check in pibridge_recv_timeout()
Instead of checking the jiffies returned by wait_event_timeout() check
directly if the requested number of bytes is available in the read queue.
This has the advantage to cover the case when the timeout specified in
wait_event_timeout() has expired before the requested data has arrived but
then the data arrives immediately after the function returns.

By doing this remove superfluous brackets in the if-statement.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>