pibridge-serdev: clear header struct before sending io message
Clear the packet header for the io message before sending the message. This
fixes a bug in which the communication with the device fails, due to an
invalid message message type (response instead of request) specified by the
response flag in the message header. The set flag is data garbage which
may be a leftover of a packet that has been received before (note that the
same data structure is also used to receive the header of response
messages).

Although this is not an issue when sending gate messages, also clear the
gate message header for consistency.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>