serial: amba-pl011: fix race condition at shutdown
In pl011_shutdown() first stopping the tx thread and then freeing the
interrupts may lead to a race condition in which the interrupt handler
is still running when the thread is stopped via kthread_stop(). Since that
function also frees the thread, the interrupt handler may access a NULL
pointer as soon as it tries to wake up the tx thread.

Fix this by first freeing the interrupt (and thus making sure that the
interrupt handler is not running any more) and then stopping the tx thread.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>