serial: Do not hold the port lock when setting rx-during-tx GPIO
Both the imx and stm32 driver set the rx-during-tx GPIO in the
rs485_config() function by means of gpiod_set_value(). Since rs485_config()
is called with the port lock held, this can be an issue in case that
setting the GPIO line can sleep (e.g. if a GPIO expander is used which is
connected via SPI or I2C).

Avoid this by setting the GPIO outside of the port lock in the serial core
and by using gpiod_set_value_cansleep() instead of gpiod_set_value().

Since now both the term and the rx-during-tx GPIO are set within the core
API use a common function uart_set_rs485_gpios() to set both.

This change also makes setting the rx-during-tx GPIO generic for all UART
drivers.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>